libshims: Properly shim VectorImpl
VectorImpl needs to have virtual methods added to the object itself.
You cannot do this by simply adding the symbols, the object itself
needs to be created with the correct sized virtual function table.

Create a separate lib that does this shimming because we need to
add to each shared library which makes use of the VectorImpl symbols
to make sure the libcutils versions aren't used.

Change-Id: I3677c55b819fe891f9d577fd6e26447c45c46555